pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "okeanos-ai-platform"
version = "0.1.0"
description = "Okeanos AI Marketing Platform MVP backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.5.0",
  "email-validator>=2.2.0",
  "sqlalchemy>=2.0.35",
  "psycopg[binary]>=3.2.1",
  "alembic>=1.13.2",
  "celery>=5.4.0",
  "redis>=5.0.8",
  "structlog>=24.4.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "httpx>=0.27.2",
  "ruff>=0.6.8",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
